requirements.txt
moondream==0.0.6
pywaggle[vision]==0.56.0
